Fix so that rename works when the item is selected.
If the identifier you want to rename is selected, the cursor
position you get from Atom actually points to the end of the word.

If you run gorename with this byte offset it will fail because the
offset actually points at the space or newline character after the
word, not the word itself.

This behavior is undesirable, as it's natural in several editors to
select a token by double clicking it and then invoking a commmand
(find, replace, rename, etc.)

We resolve this by asking Atom for the range of the currently selected
word, and then computing a position halfway between the start and end
of the word.

This works the same way whether the cursor is just sitting on the word
somewhere or if the word is actually selected.

We also grab the current identifier name and the byte offset all in the
same function now, which is slightly more efficient.